# contracts/openapi/ml-engine.yaml
openapi: 3.0.3
info:
  title: ML Engine API
  version: 1.0.0
  description: Contrato entre o Match Orchestrator e o serviço Python de inferência.
paths:
  /health:
    get:
      operationId: health
      responses:
        '200':
          description: Serviço disponível
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: UP
                  model_loaded:
                    type: boolean
                    description: Indica se o artefato job_ranker.joblib foi carregado.
                    example: true
  /predict:
    post:
      operationId: predict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRequest'
      responses:
        '200':
          description: Candidatos ordenados por compatibilidade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponse'
        '422':
          description: Payload inválido
components:
  schemas:
    Candidate:
      type: object
      required: [candidate_id, skills]
      properties:
        candidate_id:
          type: string
        skills:
          type: array
          items:
            type: string
        profession:
          type: string
        years_experience:
          type: integer
          minimum: 0
    MatchRequest:
      type: object
      required: [candidates, required_skills]
      properties:
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'
        required_skills:
          type: array
          items:
            type: string
        required_title:
          type: string
    MatchResult:
      type: object
      required: [candidate_id, score, matched_skills, skill_gaps]
      properties:
        candidate_id:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
        matched_skills:
          type: array
          items:
            type: string
        skill_gaps:
          type: array
          items:
            type: string
    MatchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/MatchResult'